test cm in nats-server tests
