+++ e2e/natsserver/natsserver_test.go
@@ -4,11 +4,13 @@
 package natsserver_test
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"os"
+	"strings"
 	"testing"
 	"time"
 
@@ -16,7 +18,11 @@ import (
 	"github.com/nats-io/nats-server/v2/server"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/kyma-project/nats-manager/api/v1alpha1"
 	. "github.com/kyma-project/nats-manager/e2e/common"
 	. "github.com/kyma-project/nats-manager/e2e/common/fixtures"
 )
@@ -25,6 +31,12 @@ const (
 	interval = 2 * time.Second
 	attempts = 60
 )
+
+// clientSet is what is used to access K8s build-in resources like Pods, Namespaces and so on.
+var clientSet *kubernetes.Clientset //nolint:gochecknoglobals // This will only be accessible in e2e tests.
+
+// k8sClient is what is used to access the NATS CR.
+var k8sClient client.Client //nolint:gochecknoglobals // This will only be accessible in e2e tests.
 
 var logger *zap.Logger
 
@@ -38,9 +50,74 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 
+	clientSet, k8sClient, err = GetK8sClients()
+	if err != nil {
+		logger.Error(err.Error())
+		panic(err)
+	}
+
+	// Update the NATS CR, so we can test some testcases.
+	ctx := context.TODO()
+	err = Retry(attempts, interval, logger, func() error {
+		cr := &v1alpha1.NATS{}
+		key := client.ObjectKey{
+			Namespace: NATSCR().GetNamespace(),
+			Name:      NATSCR().GetName(),
+		}
+		natsErr := k8sClient.Get(ctx, key, cr)
+		if natsErr != nil {
+			return natsErr
+		}
+
+		// This was disabled for earlier test. Now we activate it, so we can test against the varz-endpoint and against
+		// the configmap.
+		cr.Spec.JetStream.MemStorage.Enabled = true
+
+		natsErr = k8sClient.Update(ctx, cr)
+		return natsErr
+	})
+	if err != nil {
+		logger.Error(err.Error())
+		panic(err)
+	}
+
 	// Run the tests and exit.
 	code := m.Run()
 	os.Exit(code)
+}
+
+// Test_ConfigMap tests the ConfigMap that the NATS-Manger creates when we define a CR.
+func Test_ConfigMap(t *testing.T) {
+	ctx := context.TODO()
+
+	err := Retry(attempts, interval, logger, func() error {
+		cm, cmErr := clientSet.CoreV1().ConfigMaps(NamespaceName).Get(ctx, CMName, metav1.GetOptions{})
+		if cmErr != nil {
+			return cmErr
+		}
+
+		cmm := cmToMap(cm.Data["nats.conf"])
+
+		if err := checkValueInCMMap(cmm, "max_file", FileStorageSize); err != nil {
+			return err
+		}
+
+		if err := checkValueInCMMap(cmm, "max_mem", MemStorageSize); err != nil {
+			return err
+		}
+
+		if err := checkValueInCMMap(cmm, "debug", True); err != nil {
+			return err
+		}
+
+		if err := checkValueInCMMap(cmm, "trace", True); err != nil {
+			return err
+		}
+
+		return nil
+	})
+
+	require.NoError(t, err)
 }
 
 func Test_NATSHealth(t *testing.T) {
@@ -137,4 +214,35 @@ func getVarz(port int) (*server.Varz, error) {
 		return nil, err
 	}
 	return &varz, nil
+}
+
+func checkValueInCMMap(cmm map[string]string, key, expectedValue string) error {
+	val, ok := cmm[key]
+	if !ok {
+		return fmt.Errorf("could net get '%s' from Configmap", key)
+	}
+	if val != expectedValue {
+		return fmt.Errorf("expected value for '%s' to be '%s' but was '%s'", key, expectedValue, val)
+	}
+
+	return nil
+}
+
+func cmToMap(cm string) map[string]string {
+	lines := strings.Split(cm, "\n")
+
+	cmMap := make(map[string]string)
+	for _, line := range lines {
+		if strings.Contains(line, ": ") {
+			l := strings.Split(line, ": ")
+			if len(l) < 2 {
+				continue
+			}
+			key := strings.TrimSpace(l[0])
+			val := strings.TrimSpace(l[1])
+			cmMap[key] = val
+		}
+	}
+
+	return cmMap
 }
